requirements.txt
fastapi==0.122.0
fastapi_mail==1.5.8
Markdown==3.10
pwdlib[argon2]==0.3.0
pydantic==2.12.5
PyJWT==2.10.1
python-dotenv==1.2.1
SQLAlchemy==2.0.25
sqlmodel==0.0.27
uvicorn[standard]==0.23.2
python-multipart==0.0.6
psycopg2-binary


